# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Accounting & Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out

  # DASHBOARD
...
  /journals:
    get:
      operationId: listJournals
      tags: [journals]
      summary: List journal entries
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournal
      tags: [journals]
      summary: Create journal entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJournalBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /journals/{id}:
    get:
      operationId: getJournal
      tags: [journals]
      summary: Get journal entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteJournal
      tags: [journals]
      summary: Delete journal entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PAYMENT VOUCHERS
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payment vouchers
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create payment voucher
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete payment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # RECEIPT VOUCHERS
  /receipts:
    get:
      operationId: listReceipts
      tags: [receipts]
      summary: List receipt vouchers
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receipt"
    post:
      operationId: createReceipt
      tags: [receipts]
      summary: Create receipt voucher
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceiptBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
  /receipts/{id}:
    get:
      operationId: getReceipt
      tags: [receipts]
      summary: Get receipt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
    delete:
      operationId: deleteReceipt
      tags: [receipts]
      summary: Delete receipt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CREDIT NOTES
  /credit-notes:
    get:
      operationId: listCreditNotes
      tags: [creditNotes]
      summary: List credit notes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Credit notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditNote"
    post:
      operationId: createCreditNote
      tags: [creditNotes]
      summary: Create credit note
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditNoteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditNote"
  /credit-notes/{id}:
    get:
      operationId: getCreditNote
      tags: [creditNotes]
      summary: Get credit note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Credit note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditNote"
    delete:
      operationId: deleteCreditNote
      tags: [creditNotes]
      summary: Delete credit note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DEBIT NOTES
  /debit-notes:
    get:
      operationId: listDebitNotes
      tags: [debitNotes]
      summary: List debit notes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Debit notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DebitNote"
    post:
      operationId: createDebitNote
      tags: [debitNotes]
      summary: Create debit note
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDebitNoteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebitNote"
  /debit-notes/{id}:
    get:
      operationId: getDebitNote
      tags: [debitNotes]
      summary: Get debit note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Debit note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebitNote"
    delete:
      operationId: deleteDebitNote
      tags: [debitNotes]
      summary: Delete debit note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DELIVERY ASSIGNMENTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        stockItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        hsnCode:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        rate:
          type: number
        discountPct:
          type: number
        gstPct:
          type: number
        taxableAmount:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        total:
          type: number
      required:
        - id
        - itemName
        - quantity
        - unit
        - rate
        - discountPct
        - gstPct
        - taxableAmount
        - cgst
        - sgst
        - igst
        - total
    InvoiceItemInput:
      type: object
      properties:
        stockItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        hsnCode:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        rate:
          type: number
        discountPct:
          type: number
        gstPct:
          type: number
      required:
        - itemName
        - quantity
        - unit
        - rate
        - discountPct
        - gstPct
...
    JournalLine:
      type: object
      properties:
        id:
          type: integer
        ledgerId:
          type: integer
        ledgerName:
          type: string
        type:
          type: string
          enum: [dr, cr]
        amount:
          type: number
      required:
        - id
        - ledgerId
        - ledgerName
        - type
        - amount
    JournalLineInput:
      type: object
      properties:
        ledgerId:
          type: integer
        type:
          type: string
          enum: [dr, cr]
        amount:
          type: number
      required:
        - ledgerId
        - type
        - amount
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        voucherNumber:
          type: string
        voucherType:
          type: string
          enum: [journal, payment, receipt, contra]
        narration:
          type: string
        totalDebit:
          type: number
        totalCredit:
          type: number
        lines:
          type: array
          items:
            $ref: "#/components/schemas/JournalLine"
        createdAt:
          type: string
      required:
        - id
        - date
        - voucherNumber
        - voucherType
        - narration
        - totalDebit
        - totalCredit
        - lines
        - createdAt
    CreateJournalBody:
      type: object
      properties:
        date:
          type: string
        voucherType:
          type: string
          enum: [journal, payment, receipt, contra]
        narration:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/JournalLineInput"
      required:
        - date
        - voucherType
        - narration
        - lines
    Payment:
      type: object
      properties:
        id:
          type: integer
        voucherNumber:
          type: string
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: ["string", "null"]
        ledgerId:
          type: integer
        ledgerName:
          type: string
        paymentMode:
          type: string
        amount:
          type: number
        narration:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - voucherNumber
        - date
        - ledgerId
        - ledgerName
        - paymentMode
        - amount
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: ["string", "null"]
        ledgerId:
          type: integer
        paymentMode:
          type: string
        amount:
          type: number
        narration:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
      required:
        - date
        - ledgerId
        - paymentMode
        - amount
    Receipt:
      type: object
      properties:
        id:
          type: integer
        voucherNumber:
          type: string
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: ["string", "null"]
        ledgerId:
          type: integer
        ledgerName:
          type: string
        paymentMode:
          type: string
        amount:
          type: number
        narration:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - voucherNumber
        - date
        - ledgerId
        - ledgerName
        - paymentMode
        - amount
        - createdAt
    CreateReceiptBody:
      type: object
      properties:
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: ["string", "null"]
        ledgerId:
          type: integer
        paymentMode:
          type: string
        amount:
          type: number
        narration:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
      required:
        - date
        - ledgerId
        - paymentMode
        - amount
    CreditNote:
      type: object
      properties:
        id:
          type: integer
        noteNumber:
          type: string
        date:
          type: string
        saleInvoiceId:
          type: ["integer", "null"]
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        reason:
          type: string
        amount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
      required:
        - id
        - noteNumber
        - date
        - partyName
        - reason
        - amount
        - items
        - createdAt
    CreateCreditNoteBody:
      type: object
      properties:
        date:
          type: string
        saleInvoiceId:
          type: ["integer", "null"]
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        reason:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
      required:
        - date
        - partyName
        - reason
        - items
    DebitNote:
      type: object
      properties:
        id:
          type: integer
        noteNumber:
          type: string
        date:
          type: string
        purchaseInvoiceId:
          type: ["integer", "null"]
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        reason:
          type: string
        amount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
      required:
        - id
        - noteNumber
        - date
        - partyName
        - reason
        - amount
        - items
        - createdAt
    CreateDebitNoteBody:
      type: object
      properties:
        date:
          type: string
        purchaseInvoiceId:
          type: ["integer", "null"]
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        reason:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
      required:
        - date
        - partyName
        - reason
        - items
...